requirements.txt
anyio==4.0.0
APScheduler==3.10.1
async-timeout==4.0.3
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.10.0
Cython==3.0.0
docker==6.1.3
fastapi==0.85.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.44.0
fsspec==2023.10.0
GPUtil==1.4.0
humanfriendly==10.0
idna==2.10
imageio==2.32.0
Jinja2==3.1.2
kiwisolver==1.4.5
lapx==0.5.5
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.25.2
onnxruntime==1.15.1
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.74
packaging==23.2
pandas==2.1.2
piexif==1.1.3
Pillow==10.0.1
prometheus-client==0.18.0
prometheus-fastapi-instrumentator==6.0.0
protobuf==4.25.0
psutil==5.9.6
py-cpuinfo==9.0.0
pybase64==1.3.1
pydantic==1.10.13
Pygments==2.16.1
pyparsing==2.4.7
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.5.2
roboflow==1.1.9
scikit-image==0.22.0
scipy==1.11.3
seaborn==0.13.0
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
sympy==1.12
thop==0.1.1.post2209072238
tifffile==2023.9.26
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
websocket-client==1.6.4
